rework library loading
Loading the same named library multiple times will
result in exceptions being thrown by pywincffi when
a C-function is redefined.  To work around this the
new Library class will cache results and return the
cached value instead of calling dlopen/cdef each
time.  This change makes it possible to load a library
in multiple modules instead of having to have a central
definition of 'kernel32' that everything must import.