remove lint/doc build for Python 2.6
Python 2.6 should lint the same as other
versions of Python and the docs should build
the same as well.  Readthedocs.org does not build
using Python 2.6 anyway and the link check (which is
what's failing now) shouldn't be any different on 2.6
versus other Python versions.  Syntax wise, we're still
testing against Python 2.6 so removing the step for lint/doc
build should be ok.