HEADER_FILES, SOURCE_FILES should be read-only
Before this commit the HEADER_FILES and SOURCE_FILES
global could be and were modified by the tests.  This makes
it easy to introduce bugs into test code but it also makes it
possible for runtime modification of the headers and source.  While
technically there's nothing wrong with this the more obvious thing
is to provide arguments with modified files to the functions within dist()
instead.