simplify load()/caching and use LibraryWrapper
The old implementation of load() used a class to wrap the results
in another class.  While to an end user the results are the same
the underlying code could be cleaner.  This commit changes load()
so we're now using a couple of class methods an a simple tuple
to store the cache.