remove logging from pywincffi.core
The additional logging in pywincffi.core no longer
serves a useful purpose.  When development first
started the logging helped to track down odd bugs
in the tests but now it's just additional code calls.

In terms of testing the logging itself is turning out
to be counter-productive.  Whenever a test fails all
of the logging output ends up being shown in the test
output.  While the capturing of log output could be
disabled the log messages themselvles don't provide
a lot of useful context.