update signature of ReadFile() to match underlying API
This commit updates the signature of ReadFile() to
better match the underlying Windows API. By doing so
we should be able to better support overlapped I/O and
give the user of this function greater control over the
input and output.

For a complete discussion around why this change was made
see #141.